Add files via upload
Here, I implemented the priority queue using a 2D array and two 1D arrays named rear and front. And finally , after the end of the enqueue, dequeue and display function, I returned the array of pointers whose first element points to the first element of the rear array, whose 2nd element points to the 1st element of the front array and whose rest of the elements point to the first element of each row of the 2D array respectively.